refactor: split text editor components
